Remove td_t from sgxtypes.h
sgxtypes.h contains structures defined by the SGX specification. td_t is
a thread data structure defined by OE and is an OE implementation
detail, not an SGX one. Move td_t and related structures and functions
to their own internal header.
