+++ enclave/core/sgx/hostcalls.c
@@ -4,6 +4,7 @@
 #include <openenclave/edger8r/enclave.h>
 #include <openenclave/enclave.h>
 #include <openenclave/internal/sgx/ecall_context.h>
+#include <openenclave/internal/sgx/td.h>
 #include "td.h"
 
 /**
