+++ host/sgx/loadelf.c
@@ -13,6 +13,7 @@
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/safecrt.h>
 #include <openenclave/internal/safemath.h>
+#include <openenclave/internal/sgx/td.h>
 #include <openenclave/internal/sgxcreate.h>
 #include <openenclave/internal/trace.h>
 #include <openenclave/internal/utils.h>
